+++ packages/antdv/components/pro-modal/src/types.ts
@@ -1,7 +1,9 @@
 import type { ModalFuncProps, ModalProps } from 'ant-design-vue'
 import type { PropType } from 'vue'
 
-type ModalType = ModalFuncProps
+type ModalType = ModalFuncProps & {
+  confirmLoading?: boolean
+}
 export const ProModalProps = {
   modalProps: {
     type: Object as PropType<Omit<ModalType, 'open' | 'title'>>
@@ -28,3 +30,18 @@ export const ProModalProps = {
 //     type: Object as PropType<{ name: string }>
 //   }
 // } as const
+
+export interface ProModalSlots {
+  /**
+   * 自定义标题
+   */
+  title?: () => any
+  /**
+   * 自定义页脚
+   */
+  footer?: () => any
+  /**
+   * 默认插槽
+   */
+  default?: () => any
+}