+++ packages/antdv/components/pro-picker/src/ProPicker.vue
@@ -4,7 +4,7 @@ import { debounce } from 'lodash-es'
 import { Select, Spin } from 'ant-design-vue'
 import { reactive, watch } from 'vue'
 import { proPickerApi } from '../apis/pro-picker.api'
-import { ProPickerProps } from './types'
+import { PickerApiEnum, ProPickerProps } from './types'
 
 defineOptions({
   name: 'ProPicker'
@@ -23,9 +23,19 @@ watch(
     const type = state.data.some((item: any) => item.value === value)
     if (type)
       return
+
+    const url = PickerApiEnum[props.type]
+    proPickerApi.getOps({ url, params: { value } })
+      .then((res) => {
+        state.data = res.data.options.map((item: any) => ({
+          label: item.label,
+          value: item.value
+        }))
+      })
     if (typeof value === 'string' || typeof value === 'number') {
       proPickerApi
-        .company({
+        .getOps({
+          url,
           params: { value }
         })
         .then((res) => {
@@ -37,7 +47,8 @@ watch(
     }
     else if (Array.isArray(value)) {
       proPickerApi
-        .company({
+        .getOps({
+          url,
           params: { value: value.join(',') }
         })
         .then((res) => {
@@ -56,26 +67,15 @@ watch(
 const handleSearch = debounce((value: string) => {
   if (value) {
     state.fetching = true
-    if (props.type === 'company') {
-      proPickerApi
-        .company({
-          params: { label: value }
-        })
-        .then((res) => {
-          state.data = res.data.options.map((item: any) => ({
-            label: item.label,
-            value: item.value
-          }))
-          state.fetching = false
-        })
-    }
-    else {
-      proPickerApi
-        .user({
-          params: { label: value }
-        })
-        .then(() => {})
-    }
+    const url = PickerApiEnum[props.type]
+    proPickerApi.getOps({ url, params: { label: value } })
+      .then((res) => {
+        state.data = res.data.options.map((item: any) => ({
+          label: item.label,
+          value: item.value
+        }))
+        state.fetching = false
+      })
   }
 }, props.delay)
 </script>