+++ packages/use/index.ts
@@ -1,3 +1,4 @@
 export * from './src/hooks/file/file'
 export * from './src/hooks/Scrollbars'
 export * from './src/hooks/report/report'
+export * from './src/hooks/sms/use-sms-code'